Display OK when AT SET event if full config not ok
Everytime we set an attribute usin an AT SET command, we reload the full
configuration. While loading a check is done that the full configuration
is valid.
If some attribute is not valid, we used to display an error message.
Now we will respond with OK { attribute changed } instead.
Configuration errors will still be shown when trying to connect or when
resetting the device.